added blockstates for the screen multiblock
TODO:
The bug has something to do with how directions with checking the pattern works. I'm not sure how to fix it, and as of right now only if the keyboard block entity is facing north it works. even when checking other directions it only works facing north for some reason